train: 2026-08-24 PM (1 changes) (#112)

+++ crates/core/boss-testing/build.rs
@@ -52,11 +52,17 @@ fn main() {
     // prefix decides, and a file without one sorts last rather than
     // silently landing in the middle of the run.
     entries.sort_by_key(|n| {
-        let num: u32 = n
+        // u64, not u32: a `YYYYMMDDHHMM-` prefix is ~2.0e11 and would
+        // overflow a u32 into the no-prefix arm. That arm happens to
+        // sort timestamps last and, being a string tiebreak on equal
+        // widths, happens to keep them chronological — accidentally
+        // right, for a reason no reader could rely on. Parsed as what
+        // it is, the ordering is intent rather than luck.
+        let num: u64 = n
             .split('-')
             .next()
             .and_then(|p| p.parse().ok())
-            .unwrap_or(u32::MAX);
+            .unwrap_or(u64::MAX);
         (num, n.clone())
     });
